Move today item, add core extension hide view recycler and fix remarks
