+++ app/src/main/java/jt/projects/perfectday/presentation/today/adapter/main/FriendsViewHolder.kt
@@ -4,8 +4,8 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import jt.projects.perfectday.databinding.LayoutBithdayFriendsBinding
-import jt.projects.perfectday.presentation.today.adapter.TodayItem
 import jt.projects.perfectday.presentation.today.adapter.birth.BirthdayListAdapter
+import jt.projects.utils.extensions.hideViewInRecycler
 
 class FriendsViewHolder private constructor(
     binding: LayoutBithdayFriendsBinding
@@ -21,6 +21,10 @@ class FriendsViewHolder private constructor(
     )
 
     fun bind(item: TodayItem.Friends) {
+        if (item.friendsVk.isEmpty()) {
+            itemView.hideViewInRecycler()
+            return
+        }
         birthdayAdapter.submitList(item.friendsVk)
     }
 }
