+++ app/src/main/java/jt/projects/perfectday/presentation/today/adapter/main/NoteViewHolder.kt
@@ -4,7 +4,6 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import jt.projects.perfectday.databinding.ItemScheduledEventBinding
-import jt.projects.perfectday.presentation.today.adapter.TodayItem
 import jt.projects.utils.toStdFormatString
 
 class NoteViewHolder private constructor(
