+++ app/src/main/java/jt/projects/perfectday/presentation/today/adapter/main/TodayDiffUtil.kt
@@ -1,7 +1,6 @@
 package jt.projects.perfectday.presentation.today.adapter.main
 
 import androidx.recyclerview.widget.DiffUtil
-import jt.projects.perfectday.presentation.today.adapter.TodayItem
 
 class TodayDiffUtil : DiffUtil.ItemCallback<TodayItem>() {
     override fun areItemsTheSame(oldItem: TodayItem, newItem: TodayItem): Boolean =
