+++ app/src/main/res/layout/item_fact_of_day.xml
@@ -5,7 +5,6 @@
     android:layout_height="wrap_content"
     android:background="@drawable/bg_setting"
     android:orientation="vertical"
-    android:layout_marginTop="25dp"
     android:paddingHorizontal="12dp"
     android:paddingVertical="20dp">
